library: fix authentication issue while reconnecting
After a device disconnects and tries to connect again, the service becomes unresponsible
as it still believes the device is authenticated. This commit forces the authentication
manager to re-authenticate if a new valid authentication request is received. This change
is not harmful as only one active connection is supported at once.

Signed-off-by: David Escalona <david.escalona@digi.com>