library: peripheral: fix peripheral component to make it work with low MTU rates
The peripheral component was developed to work with high MTU rates (<250). This means that the received
messages were expected to be single and complete atomic XBee frames. Same applied to the data to send,
it was sent in just one operation.

This commit modifies the peripheral component to work with low MTU rates as follows:
- Modify the read data callback to consider received data as complete XBee frames or just a chunk of a
  frame. This is done by saving received data in a buffer until the complete message is received.
- Modify the write data method to slice the data to send as necessary depending on the negotiated MTU.
  Since the MTU cannot be directly determined by the library, start with a conservative MTU (27 -3 bytes)
  and consider the received data lenght to increase the MTU accordingly.

Signed-off-by: David Escalona <david.escalona@digi.com>